backend/pyproject.toml
[project]
name = "nanowebui"
version = "0.1.0"
description = "Web UI for managing nanobot AI agent"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "psutil>=6.0.0",
    "websockets>=13.0",
    "httpx>=0.27.0",
]

[project.scripts]
nanowebui = "app.main:start"
